requirements.txt
psutil
Flask
requests
openai
python-dotenv
Pillow
PyMuPDF